:recycle:[refactor]: 함수들 모듈화하여 특징별 파일 분류
